feat: reorganização das pastas
