+++ front-end/trabalho-dac/src/app/pages/cliente/menu-cliente/menu-cliente.component.spec.ts
@@ -1,0 +1,23 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { MenuClienteComponent } from './menu-cliente.component';
+
+describe('MenuClienteComponent', () => {
+  let component: MenuClienteComponent;
+  let fixture: ComponentFixture<MenuClienteComponent>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      imports: [MenuClienteComponent]
+    })
+    .compileComponents();
+    
+    fixture = TestBed.createComponent(MenuClienteComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
