+++ front-end/trabalho-dac/src/app/pages/cliente/modal-cancelar/modal-cancelar.component.ts
@@ -1,7 +1,7 @@
 import { Component, Input } from '@angular/core';
 import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';
-import { ReservaService } from '../../services/reserva.service';
-import { UserService } from '../../services/user.service';
+import { ReservaService } from '../../../services/reserva.service';
+import { UserService } from '../../../services/user.service';
 
 @Component({
   selector: 'app-modal-cancelar',
